+++ src/diagnostic/DiagnosticPane.js
@@ -6,18 +6,6 @@ import './DiagnosticPane.css';
 import GeneratePackage from './GeneratePackage';
 import Ping from './Ping';
 import uuid from 'uuid';
-// import JMXDisplay from '../jmx/JMXDisplay';
-// import { Cypher } from 'graph-app-kit/components/Cypher';
-
-// const renderJMX = ({ pending, error, result }) => {
-//     return pending ? (
-//       <div style={{ height: "60px" }}>pending</div>
-//     ) : error ? (
-//       <div style={{ height: "60px" }}>{error.message}</div>
-//     ) : result ? (
-//       <JMXDisplay data={result} />
-//     ) : null;
-//   };
 
 class DiagnosticPane extends Component {
     state = {
@@ -31,12 +19,6 @@ class DiagnosticPane extends Component {
                     <Grid.Row columns={1}>
                         <Grid.Column>
                             <Overview key={this.state.key} node={this.props.node} driver={this.props.driver}/>
-                        </Grid.Column>
-                    </Grid.Row>
-
-                    <Grid.Row columns={1}>
-                        <Grid.Column>
-                            <Ping key={this.state.key} node={this.props.node} driver={this.props.driver}/>
                         </Grid.Column>
                     </Grid.Row>
 
@@ -56,11 +38,11 @@ class DiagnosticPane extends Component {
                         </Grid.Column>
                     </Grid.Row>
 
-                    {/* <Grid.Row columns={1}>
+                    <Grid.Row columns={1}>
                         <Grid.Column>
-                            <Cypher query="CALL dbms.queryJmx('*:*')" render={renderJMX} interval={3000} />
+                            <Ping key={this.state.key} node={this.props.node} driver={this.props.driver}/>
                         </Grid.Column>
-                    </Grid.Row> */}
+                    </Grid.Row>
                 </Grid>  
             </div>
         );