+++ src/diagnostic/GeneratePackage.js
@@ -1,11 +1,12 @@
 import React, { Component } from 'react';
 import * as PropTypes from "prop-types";
-import { Button, Icon } from 'semantic-ui-react';
+import { Button, Icon, Tab } from 'semantic-ui-react';
 import Spinner from '../Spinner';
 import uuid from 'uuid';
 import status from '../status/index';
 import moment from 'moment';
 import Advisor from './advisor/Advisor';
+import ConfigurationDiff from './ConfigurationDiff';
 import advisor from './advisor/index';
 
 class GeneratePackage extends Component {
@@ -74,16 +75,53 @@ class GeneratePackage extends Component {
             URL.createObjectURL(blob);
     };
 
+    renderDiagnosticAdvice() {
+        if (!this.state.diagnosticData) {
+            return '';
+        }
+
+        const panes = [
+            { 
+                menuItem: 'Advisor', 
+                render: () => 
+                    <Tab.Pane>
+                        <Advisor 
+                            key={uuid.v4()} 
+                            data={advisor.generateRecommendations(this.state.diagnosticData)}
+                        />
+                    </Tab.Pane>,
+            },
+            { 
+                menuItem: 'Configuration Diff', 
+                render: () => 
+                    <Tab.Pane>
+                        <ConfigurationDiff data={this.state.diagnosticData} />
+                    </Tab.Pane> 
+            },
+        ];
+
+        return (<Tab menu={{ borderless: true, attached: false, tabular: false }} panes={panes} />);
+    }
+
     render() {
         let message = status.formatStatusMessage(this);
 
         return (
             <div className='GeneratePackage'>
-                <Button basic
+                <Button basic disabled={this.state.loading}
                         onClick={this.generatePackage}>
                     <Icon name='cogs'/>
                     Run Diagnostics!
                 </Button>
+
+                { this.state.diagnosticData ? (
+                    <Button basic 
+                        download={`neo4j-diagnostics-${this.state.dataGenerated}.json`}
+                        href={this.buildURI(this.state.diagnosticData)}>
+                        <Icon name="download"/>
+                        Download Diagnostics
+                    </Button>
+                ) : '' }
 
                 <div style={{
                     marginTop: '15px',
@@ -98,21 +136,7 @@ class GeneratePackage extends Component {
                     this.state.loading ? <Spinner active={this.state.loading} /> : ''
                 }
 
-                { this.state.diagnosticData ? 
-                    <Advisor 
-                        key={uuid.v4()} 
-                        data={advisor.generateRecommendations(this.state.diagnosticData)}
-                    /> : 
-                  '' }
-
-                { this.state.diagnosticData ? (
-                    <Button basic 
-                        download={`neo4j-diagnostics-${this.state.dataGenerated}.json`}
-                        href={this.buildURI(this.state.diagnosticData)}>
-                        <Icon name="download"/>
-                        Download Diagnostics
-                    </Button>
-                ) : '' }
+                { this.renderDiagnosticAdvice() }
             </div>
         );
     }