+++ src/overview/ClusterOverviewPane.js
@@ -1,0 +1,50 @@
+import React, { Component } from 'react';
+import { Grid } from 'semantic-ui-react';
+
+import ClusterMemory from './ClusterMemory';
+import GCPauseTime from './GCPauseTime';
+// import PageCacheFaults from './PageCacheFaults';
+// import PageCacheFlushes from './PageCacheFlushes';
+import PageCacheTracking from './PageCacheTracking';
+import TransactionsOpen from './TransactionsOpen';
+import UsedMemory from './UsedMemory';
+import OpenFileDescriptors from './OpenFileDescriptors';
+
+export default class ClusterOverviewPane extends Component {
+    render() {
+        return (
+            <div className="PerformancePane">
+                <Grid divided='vertically'>
+                    <Grid.Row columns={3}>
+                        <Grid.Column>
+                            <ClusterMemory />
+                        </Grid.Column>
+
+                        <Grid.Column>
+                            <UsedMemory />
+                        </Grid.Column>
+
+                        <Grid.Column>
+                            <GCPauseTime />
+                        </Grid.Column>
+                    </Grid.Row>
+
+                    <Grid.Row columns={3}>
+                        <Grid.Column>
+                            <PageCacheTracking />
+                        </Grid.Column>
+
+                        <Grid.Column>
+                            <TransactionsOpen />
+                        </Grid.Column>
+
+                        <Grid.Column>
+                            <OpenFileDescriptors />
+                        </Grid.Column>
+                    </Grid.Row>
+                </Grid>
+            </div>
+        );
+    }
+};
+