+++ src/performance/OSPane.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import "semantic-ui-css/semantic.min.css";
 import { Grid } from 'semantic-ui-react';
-// import HeapComponent from './HeapComponent';
 import DiskUsage from './DiskUsage';
 import PageCache from '../diagnostic/PageCache';
 import OSStats from './OSStats';