+++ src/performance/TransactionMonitor.js
@@ -3,30 +3,37 @@ import CypherTimeseries from '../timeseries/CypherTimeseries';
 import uuid from 'uuid';
 import queryLibrary from '../data/query-library';
 
-class MemoryMonitor extends Component {
+class TransactionMonitor extends Component {
     state = {
         key: uuid.v4(),
         rate: 1000,
         width: 400,
-        query: queryLibrary.JMX_MEMORY_STATS.query,
-        displayColumns: queryLibrary.JMX_MEMORY_STATS.columns,
+        query: queryLibrary.JMX_TRANSACTIONS.query,
+        displayColumns: queryLibrary.JMX_TRANSACTIONS.columns,
+        legendOnlyColumns: queryLibrary.JMX_TRANSACTIONS.legendOnlyColumns,
+    };
+
+    onUpdate = (childQueryState) => {
+        // console.log('child query state',childQueryState);
     };
 
     render() {
         return (
-            <div className="MemoryMonitor">
+            <div className="TransactionMonitor">
                 <CypherTimeseries key={this.state.key}
                     driver={this.props.driver}
                     node={this.props.node}
                     query={this.state.query} 
                     width={this.state.width}
                     rate={this.state.rate}
+                    onUpdate={this.onUpdate}
                     startingEnabled={[this.state.displayColumns[1]]}
                     displayColumns={this.state.displayColumns}
+                    legendOnlyColumns={this.state.legendOnlyColumns}
                 />
             </div>
         )
     }
 }
 
-export default MemoryMonitor;
+export default TransactionMonitor;